ServerManagementJob/pyproject.toml
[tool.poetry]
name = "server-mgmt-job"
version = "0.1.0"
description = ""
authors = ["Elusive Zenith <26206679+ElusiveZenith@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.140"
requests = "^2.32.3"
construct = "^2.10.70"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
